requirements.txt
asgiref==3.3.4
astroid==2.5.3
autopep8==1.5.6
colorama==0.4.4
Django==3.0.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
